vn(feat) gh auth
