+++  .env.local
@@ -1,0 +1,3 @@
+# .env.local
+
+NEXTAUTH_URL="localhost:3000/api/rm/v0/auth"
