+++ lib/auth/index.ts
@@ -1,0 +1,2 @@
+// index.ts
+export { authOptions } from "./constants";
