--- src/app/api/rm/v0/auth/signin.ts
@@ -1,39 +1,0 @@
-// [...nextauth].ts// auth.ts TS-Doc?
-
-import NextAuth from "next-auth";
-import GithubProvider from "next-auth/providers/github";
-import EmailProvider from "next-auth/providers/email";
-import { MongoDBAdapter } from "@next-auth/mongodb-adapter";
-import { MongoConnector } from "@model";
-
-export const authOptions = {
-  // Configure one or more authentication providers
-  providers: [
-    GithubProvider({
-      clientId: process.env.GITHUB_ID as string,
-      clientSecret: process.env.GITHUB_SECRET as string,
-    }),
-    EmailProvider({
-      server: process.env.EMAIL_SERVER as string,
-      from: process.env.EMAIL_FROM as string,
-      // maxAge: 24 * 60 * 60, // How long email links are valid for (default 24h)
-    }),
-    // ...add more providers here
-  ],
-  pages: {
-    // signIn: '/auth/signin',
-    // signOut: '/auth/signout',
-    // error: '/auth/error', // Error code passed in query string as ?error=
-    // verifyRequest: '/auth/verify-request', // (used for check email message)
-    // newUser: '/auth/new-user' // New users will be directed here on first sign in (leave the property out if not of interest)
-  },
-};
-
-// For more information on each option (and a full list of options) go to
-// https://authjs.dev/reference/providers/oauth
-const handler = NextAuth({
-  adapter: MongoDBAdapter(MongoConnector),
-  ...authOptions,
-});
-
-export { handler as GET, handler as POST };
