+++ tsconfig.json
@@ -22,7 +22,8 @@
       "@styles/*": ["./styles/*"],
       "@model": ["./lib/model"],
       "@model/*": ["./lib/model/*"],
-      "@auth/*": ["./src/app/api/rm/v0/auth/*"],
+      "@auth": ["./lib/auth"],
+      "@auth/*": ["./lib/auth/*"],
     },
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
